Rename draw-to-insert-strategy.tsx file
