+++ editor/src/components/canvas/canvas-strategies/strategies/draw-to-insert-metastrategy.tsx
